+++ app/src/main/kotlin/com/kelsos/mbrc/ui/dialogs/OutputSelectionDialog.kt
@@ -112,7 +112,7 @@ class OutputSelectionDialog : DialogFragment(),
     availableOutputs.show()
   }
 
-  override fun error(@OutputSelectionContract.Code code: Long) {
+  override fun error(@OutputSelectionContract.Code code: Int) {
     val resId = when (code) {
       OutputSelectionContract.CONNECTION_ERROR -> R.string.output_selection__connection_error
       else -> R.string.output_selection__generic_error
